+++ lib/presentation/providers/annuity/annuity_form_provider.dart
@@ -4,21 +4,18 @@ import 'package:ingemath/domain/domain.dart';
 import 'package:ingemath/infraestructure/infraestructure.dart';
 import 'package:ingemath/presentation/providers/providers.dart';
 
-final annuityFormProvider = StateNotifierProvider
-  .autoDispose<AnnuityFormNotifier, AnnuityFormState>((ref) {
-
-    final annuityRepository = ref.watch(annuityRepositorytProvider);
-    return AnnuityFormNotifier(
-      repository: annuityRepository,
-    );
-
-  }
-);
+final annuityFormProvider =
+    StateNotifierProvider.autoDispose<AnnuityFormNotifier, AnnuityFormState>(
+        (ref) {
+  final annuityRepository = ref.watch(annuityRepositorytProvider);
+  return AnnuityFormNotifier(
+    repository: annuityRepository,
+  );
+});
 
 enum AnnuityVariable { none, amount, annuityValue, interestRate, time }
 
 class AnnuityFormState {
-
   final menuOptions = const <AnnuityVariable, String>{
     AnnuityVariable.amount: "Monto anualidad",
     AnnuityVariable.annuityValue: "Valor actual de la anualidad",
@@ -55,17 +52,17 @@ class AnnuityFormState {
     InterestRate? interestRate,
     DataNumber? time,
     double? result,
-  }) => AnnuityFormState(
-
-    isFormPosted: isFormPosted ?? this.isFormPosted,
-    isValid: isValid ?? this.isValid,
-    variable: variable ?? this.variable,
-    amount: amount ?? this.amount,
-    annuityValue: annuityValue ?? this.annuityValue,
-    interestRate: interestRate ?? this.interestRate,
-    time: time ?? this.time,
-    result: result ?? this.result,
-  );
+  }) =>
+      AnnuityFormState(
+        isFormPosted: isFormPosted ?? this.isFormPosted,
+        isValid: isValid ?? this.isValid,
+        variable: variable ?? this.variable,
+        amount: amount ?? this.amount,
+        annuityValue: annuityValue ?? this.annuityValue,
+        interestRate: interestRate ?? this.interestRate,
+        time: time ?? this.time,
+        result: result ?? this.result,
+      );
 }
 
 class AnnuityFormNotifier extends StateNotifier<AnnuityFormState> {
@@ -93,7 +90,7 @@ class AnnuityFormNotifier extends StateNotifier<AnnuityFormState> {
     );
   }
 
-  void onInterestRateChanged(int value) {
+  void onInterestRateChanged(double value) {
     state = state.copyWith(
       interestRate: InterestRate.dirty(value),
     );
@@ -111,8 +108,8 @@ class AnnuityFormNotifier extends StateNotifier<AnnuityFormState> {
     if (!state.isValid) return;
 
     double result = 0;
-    switch(state.variable) {
-      case AnnuityVariable.amount :
+    switch (state.variable) {
+      case AnnuityVariable.amount:
         result = await repository.calculateAmount(
           interestRate: state.interestRate.value,
           annuityValue: state.annuityValue.value,
@@ -120,7 +117,7 @@ class AnnuityFormNotifier extends StateNotifier<AnnuityFormState> {
         );
         break;
 
-      case AnnuityVariable.annuityValue :
+      case AnnuityVariable.annuityValue:
         result = await repository.calculateAnnuityValue(
           interestRate: state.interestRate.value,
           amount: state.amount.value,
@@ -128,22 +125,22 @@ class AnnuityFormNotifier extends StateNotifier<AnnuityFormState> {
         );
         break;
 
-      case AnnuityVariable.interestRate :
+      case AnnuityVariable.interestRate:
         result = await repository.calculateInterestRate(
           amount: state.amount.value,
           annuityValue: state.annuityValue.value,
           time: state.time.value,
         );
         break;
-        
-      case AnnuityVariable.time :
+
+      case AnnuityVariable.time:
         result = await repository.calculateTime(
           amount: state.amount.value,
           annuityValue: state.annuityValue.value,
           interestRate: state.interestRate.value,
         );
         break;
-      default :
+      default:
         break;
     }
     state = state.copyWith(result: result);
@@ -151,19 +148,20 @@ class AnnuityFormNotifier extends StateNotifier<AnnuityFormState> {
 
   void _touchEveryField() {
     state = state.copyWith(
-
       isFormPosted: true,
       amount: DataNumber.dirty(state.amount.value),
       annuityValue: DataNumber.dirty(state.annuityValue.value),
       time: DataNumber.dirty(state.time.value),
       interestRate: InterestRate.dirty(state.interestRate.value),
-
-      isValid: state.variable != AnnuityVariable.none && Formz.validate([
-        if(state.variable != AnnuityVariable.amount) state.amount,
-        if(state.variable != AnnuityVariable.annuityValue) state.annuityValue,
-        if(state.variable != AnnuityVariable.interestRate) state.interestRate,
-        if(state.variable != AnnuityVariable.time) state.time,
-      ]),
+      isValid: state.variable != AnnuityVariable.none &&
+          Formz.validate([
+            if (state.variable != AnnuityVariable.amount) state.amount,
+            if (state.variable != AnnuityVariable.annuityValue)
+              state.annuityValue,
+            if (state.variable != AnnuityVariable.interestRate)
+              state.interestRate,
+            if (state.variable != AnnuityVariable.time) state.time,
+          ]),
     );
   }
 }
